Move types and tests for hunk-dependencies to `but-hunk-dependency`
The UI is conceptionally something that the backend talks via JSON based RPC.
