+++ crates/but-hunk-dependency/Cargo.toml
@@ -22,11 +22,14 @@ git2.workspace = true
 gitbutler-repo.workspace = true
 gitbutler-oxidize.workspace = true
 
+# For `ui` module
+rustc-hash = "2.1.0"
+gitbutler-stack.workspace = true
+
 [dev-dependencies]
 gix-testtools.workspace = true
-gitbutler-stack.workspace = true
 gitbutler-testsupport.workspace = true
-gitbutler-command-context.workspace = true
 gitbutler-oxidize.workspace = true
 itertools = "0.14.0"
 insta.workspace = true
+serde_json = "1.0.138"
