+++ crates/gitbutler-tauri/tests/tauri/main.rs
@@ -1,2 +1,1 @@
 mod diff;
-mod workspace;
